refactor: split cho domain
