+++ app/api/domains/__init__.py
@@ -2,4 +2,3 @@
 
 from . import cho
 from . import map
-from . import osu
