+++ app/api/rest/v2/models/clans.py
