+++ app/api/web/beatmaps.py
@@ -1,0 +1,99 @@
+from __future__ import annotations
+
+from fastapi import status
+from fastapi.param_functions import Depends
+from fastapi.param_functions import Header
+from fastapi.param_functions import Query
+from fastapi.requests import Request
+from fastapi.responses import RedirectResponse
+from fastapi.responses import Response
+from fastapi.routing import APIRouter
+
+import app.state
+from app.api.web.authentication import authenticate_player_session
+from app.logging import Ansi
+from app.logging import log
+from app.objects import models
+from app.objects.player import Player
+from app.objects.score import SubmissionStatus
+from app.repositories import maps as maps_repo
+from app.repositories import scores as scores_repo
+
+router = APIRouter()
+
+
+def bancho_to_osuapi_status(bancho_status: int) -> int:
+    return {
+        0: 0,
+        2: 1,
+        3: 2,
+        4: 3,
+        5: 4,
+    }[bancho_status]
+
+
+@router.post("/osu-getbeatmapinfo.php")
+async def osuGetBeatmapInfo(
+    form_data: models.OsuBeatmapRequestForm,
+    player: Player = Depends(authenticate_player_session(Query, "u", "h")),
+) -> Response:
+    num_requests = len(form_data.Filenames) + len(form_data.Ids)
+    log(f"{player} requested info for {num_requests} maps.", Ansi.LCYAN)
+
+    response_lines: list[str] = []
+
+    for idx, map_filename in enumerate(form_data.Filenames):
+        # try getting the map from sql
+
+        beatmap = await maps_repo.fetch_one(filename=map_filename)
+
+        if not beatmap:
+            continue
+
+        # try to get the user's grades on the map
+        # NOTE: osu! only allows us to send back one per gamemode,
+        #       so we've decided to send back *vanilla* grades.
+        #       (in theory we could make this user-customizable)
+        grades = ["N", "N", "N", "N"]
+
+        for score in await scores_repo.fetch_many(
+            map_md5=beatmap["md5"],
+            user_id=player.id,
+            mode=player.status.mode.as_vanilla,
+            status=SubmissionStatus.BEST,
+        ):
+            grades[score["mode"]] = score["grade"]
+
+        response_lines.append(
+            "{i}|{id}|{set_id}|{md5}|{status}|{grades}".format(
+                i=idx,
+                id=beatmap["id"],
+                set_id=beatmap["set_id"],
+                md5=beatmap["md5"],
+                status=bancho_to_osuapi_status(beatmap["status"]),
+                grades="|".join(grades),
+            ),
+        )
+
+    if form_data.Ids:  # still have yet to see this used
+        await app.state.services.log_strange_occurrence(
+            f"{player} requested map(s) info by id ({form_data.Ids})",
+        )
+
+    return Response("\n".join(response_lines).encode())
+
+
+@router.get("/maps/{map_filename}")
+async def get_updated_beatmap(
+    request: Request,
+    map_filename: str,
+    host: str = Header(...),
+) -> Response:
+    """Send the latest .osu file the server has for a given map."""
+    if host == "osu.ppy.sh":
+        return Response("bancho.py only supports the -devserver connection method")
+
+    return RedirectResponse(
+        url=f"https://osu.ppy.sh{request['raw_path'].decode()}",
+        status_code=status.HTTP_301_MOVED_PERMANENTLY,
+    )
