+++ app/api/web/connect.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+from fastapi.param_functions import Query
+from fastapi.responses import Response
+from fastapi.routing import APIRouter
+
+router = APIRouter()
+
+
+@router.get("/bancho_connect.php")
+async def banchoConnect(
+    # NOTE: this is disabled as this endpoint can be called
+    #       before a player has been granted a session
+    # player: Player = Depends(authenticate_player_session(Query, "u", "h")),
+    osu_ver: str = Query(..., alias="v"),
+    active_endpoint: str | None = Query(None, alias="fail"),
+    net_framework_vers: str | None = Query(None, alias="fx"),  # delimited by |
+    client_hash: str | None = Query(None, alias="ch"),
+    retrying: bool | None = Query(None, alias="retry"),  # '0' or '1'
+) -> Response:
+    return Response(b"")
